requirements.txt
# Core AI / Embeddings
sentence-transformers
torch
transformers

# Vector Database
faiss-cpu
numpy

# Progress bars
tqdm

# PDF Processing
pymupdf
pdf2image
pillow

# OCR
pytesseract

# Data handling
pandas
scikit-learn

# Utilities
python-dotenv

requests
streamlit